compile-and-run-a-node: various minor edits and simplifications
First of, add `--download-config` argument to `neard init`.  This
removes the ‘Replacing the config.json’ stop (since neard already gets
the proper configuration file) and makes mention of boot nodes
unnecessary (since boot nodes are listed in downloaded file).

Secondly, better document getting data backup including mentioning
it’s an optional step and that around 500GB of SSD is required to be
able to run and keep up with the network.

Fixes: https://github.com/near/node-docs/issues/37